+++ generators/htmllog.py
@@ -32,24 +32,7 @@ import logparser
 def main() -> None:
     """Generate an html log."""
 
-    try:
-        filename = pathlib.Path(sys.argv[1])
-        log = logparser.logfile(str(filename))
-    except IndexError:
-        sys.stderr.write(
-            f"""\
-\x1B[1mUsage: {sys.argv[0]} <logfile>\x1B[0m
-"""
-        )
-        sys.exit(1)
-    except OSError:
-        sys.stderr.write(
-            f"""\
-\x1B[31mopen failed!\x1B[0m
-\x1B[1mUsage: {sys.argv[0]} <logfile>\x1B[0m
-"""
-        )
-        sys.exit(1)
+    log = logparser.from_argv()
 
     # pylint: disable=too-many-return-statements
     def gen_html(ev: logparser.LogEvent) -> str:
@@ -163,6 +146,7 @@ def main() -> None:
             or ev.type == ["tbot", "info"]
             or ev.type[0] == "custom"
             or ev.type[0] == "doc"
+            or ev.type[0] == "__debug__"
         ):
             return ""
 
@@ -172,7 +156,7 @@ def main() -> None:
         template_string = f.read()
 
     data = {
-        "page_title": f"tbot log: {pathlib.Path(filename.stem)}",
+        "page_title": f"tbot log: {pathlib.Path(pathlib.Path(sys.argv[1]).stem)}",
         "content": "\n".join(map(gen_html, log)),
     }