+++ tbot/tc/selftest/tc/__init__.py
@@ -10,7 +10,7 @@ from .build import *  # noqa: F403
 
 @tbot.testcase
 def selftest_tc(lab: typing.Optional[linux.LabHost] = None,) -> None:
-    """Test builtin testcases."""
+    """Run selftests for builtin testcases."""
     with lab or tbot.acquire_lab() as lh:
         tc.testsuite(
             selftest_tc_git_checkout,  # noqa: F405